MT#55283 Added decoder mix "slots" concept ...
... so that the desired wav channel can be controlled when producing a
mixed audio file

When a mixed wav file is created, the channels in the wav container are
currently allocated in the same order as each SSRC is received, meaning
it is impossible to know which channels have been allocated to the offer
or answer side of the call. Furthermore if there is a reinvite or media
file played, these are also allocated in the order that SSRC is received
- so an "answer" could end up sharing a channel with an "offer" with no
way of knowing this.

This patch allows you to specify how many channel slots should be
allocated within the mixer, and allows you to then specify which slot is
assigned to each media in the call (this will usually be 2 slots in
total, slot 1 for answer, slot 2 for offer or vice versa).

Ported from https://github.com/sipwise/rtpengine/pull/1852

Closes #1857
Closes #1852

Change-Id: I010208427cabc3a48d6ef7bd3a84e9a5bdcfd492